+++ src/main/java/com/study/petory/domain/tradeboard/dto/request/TradeBoardCreateRequestDto.java
@@ -1,9 +1,7 @@
-package com.study.petory.domain.tradeBoard.dto.request;
+package com.study.petory.domain.tradeboard.dto.request;
 
-import com.study.petory.domain.tradeBoard.entity.TradeBoardStatus;
-import com.study.petory.domain.tradeBoard.entity.TradeCategory;
+import com.study.petory.domain.tradeboard.entity.TradeCategory;
 
-import jakarta.annotation.Nullable;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
@@ -12,7 +10,7 @@ import lombok.Getter;
 
 @Getter
 @AllArgsConstructor
-public class TradeBoardUpdateRequestDto {
+public class TradeBoardCreateRequestDto {
 
 	@NotNull(message = "카테고리를 입력해주세요.")
 	private final TradeCategory category;
@@ -27,7 +25,4 @@ public class TradeBoardUpdateRequestDto {
 
 	@NotNull(message = "금액을 입력해주세요.")
 	private final Integer price;
-
-	@NotNull(message = "판매하시는 물품의 상태를 입력해주세요.")
-	private final TradeBoardStatus status;
 }
