+++ src/main/java/com/study/petory/domain/tradeboard/dto/response/TradeBoardGetResponseDto.java
@@ -1,17 +1,19 @@
-package com.study.petory.domain.tradeBoard.dto.response;
+package com.study.petory.domain.tradeboard.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.study.petory.domain.tradeBoard.entity.TradeBoard;
-import com.study.petory.domain.tradeBoard.entity.TradeCategory;
+import com.study.petory.domain.tradeboard.entity.TradeBoard;
+import com.study.petory.domain.tradeboard.entity.TradeCategory;
 
 import lombok.Getter;
 
 @Getter
-public class TradeBoardCreateResponseDto {
+public class TradeBoardGetResponseDto {
 
 	private final Long id;
+
+	private final Long sellerId;
 
 	private final TradeCategory category;
 
@@ -19,19 +21,24 @@ public class TradeBoardCreateResponseDto {
 
 	private final String content;
 
-	private final List<String> imageUrls;
+	private final List<String> urls;
 
 	private final Integer price;
 
 	private final LocalDateTime createdAt;
 
-	public TradeBoardCreateResponseDto(TradeBoard tradeBoard, List<String> imageUrls) {
+	private final LocalDateTime updatedAt;
+
+	public TradeBoardGetResponseDto(TradeBoard tradeBoard, List<String> urls) {
 		this.id = tradeBoard.getId();
+		this.sellerId = tradeBoard.getUser().getId();
 		this.category = tradeBoard.getCategory();
 		this.title = tradeBoard.getTitle();
 		this.content = tradeBoard.getContent();
-		this.imageUrls = imageUrls;
+		this.urls = urls;
 		this.price = tradeBoard.getPrice();
 		this.createdAt = tradeBoard.getCreatedAt();
+		this.updatedAt = tradeBoard.getUpdatedAt();
 	}
+
 }
