+++ src/main/java/com/study/petory/domain/tradeboard/dto/response/TradeBoardUpdateResponseDto.java
@@ -1,20 +1,17 @@
-package com.study.petory.domain.tradeBoard.dto.response;
+package com.study.petory.domain.tradeboard.dto.response;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
-import com.study.petory.domain.tradeBoard.entity.TradeBoard;
-import com.study.petory.domain.tradeBoard.entity.TradeBoardImage;
-import com.study.petory.domain.tradeBoard.entity.TradeCategory;
+import com.study.petory.domain.tradeboard.entity.TradeBoard;
+import com.study.petory.domain.tradeboard.entity.TradeBoardStatus;
+import com.study.petory.domain.tradeboard.entity.TradeCategory;
 
 import lombok.Getter;
 
 @Getter
-public class TradeBoardGetResponseDto {
+public class TradeBoardUpdateResponseDto {
 
 	private final Long id;
-
-	private final Long sellerId;
 
 	private final TradeCategory category;
 
@@ -22,24 +19,22 @@ public class TradeBoardGetResponseDto {
 
 	private final String content;
 
-	private final List<String> urls;
-
 	private final Integer price;
+
+	private final TradeBoardStatus status;
 
 	private final LocalDateTime createdAt;
 
 	private final LocalDateTime updatedAt;
 
-	public TradeBoardGetResponseDto(TradeBoard tradeBoard, List<String> urls) {
+	public TradeBoardUpdateResponseDto(TradeBoard tradeBoard) {
 		this.id = tradeBoard.getId();
-		this.sellerId = tradeBoard.getUser().getId();
 		this.category = tradeBoard.getCategory();
 		this.title = tradeBoard.getTitle();
 		this.content = tradeBoard.getContent();
-		this.urls = urls;
 		this.price = tradeBoard.getPrice();
+		this.status = tradeBoard.getStatus();
 		this.createdAt = tradeBoard.getCreatedAt();
 		this.updatedAt = tradeBoard.getUpdatedAt();
 	}
-
 }
