+++ src/main/java/com/study/petory/domain/tradeboard/repository/TradeBoardImageRepository.java
@@ -1,8 +1,8 @@
-package com.study.petory.domain.tradeBoard.repository;
+package com.study.petory.domain.tradeboard.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.study.petory.domain.tradeBoard.entity.TradeBoardImage;
+import com.study.petory.domain.tradeboard.entity.TradeBoardImage;
 
 public interface TradeBoardImageRepository extends JpaRepository<TradeBoardImage, Long> {
 }
