+++ src/main/java/com/study/petory/domain/tradeboard/repository/TradeBoardRepository.java
@@ -1,11 +1,10 @@
-package com.study.petory.domain.tradeBoard.repository;
+package com.study.petory.domain.tradeboard.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.study.petory.domain.tradeBoard.entity.TradeBoard;
-import com.study.petory.domain.tradeBoard.entity.TradeCategory;
+import com.study.petory.domain.tradeboard.entity.TradeBoard;
 import com.study.petory.domain.user.entity.User;
 
 public interface TradeBoardRepository extends JpaRepository<TradeBoard, Long>, TradeBoardQueryRepository {
